AgenticRAG/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
tavily-python
chromadb
pypdf
python-multipart
aiofiles
httpx
tiktoken
